+++ src/App.scss
@@ -15,7 +15,7 @@
 
 .App-header {
   padding: 0 30px;
-  background-color: #37626d;
+
   min-height: 100vh;
   display: flex;
   align-items: center;
@@ -24,3 +24,23 @@
   color: white;
 }
 
+.nav {
+  position: relative;
+  border: 4px solid rgba(67, 122, 177, 0.333);
+  padding: 20px;
+  background-color: rgb(146, 218, 157);
+  &__list {
+    display: flex;
+    gap: 40px;
+  }
+}
+
+* {
+  margin: 0;
+}
+li {
+  list-style: none;
+}
+body {
+  background-color: #427e6c;
+}
