+++ src/components/ChatList/ChatList.scss
@@ -1,0 +1,10 @@
+a {
+  color: inherit;
+  text-decoration: none;
+}
+.chatList {
+  background-color: #427e6c;
+}
+.css-10hburv-MuiTypography-root {
+  color: aliceblue;
+}
