+++ src/components/Profile/Profile.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export function Profile() {
+  return (
+    <div>Profile</div>
+  )
+}
