+++ src/components/form/Form.scss
@@ -47,12 +47,12 @@
 }
 .form-content {
   display: flex;
-  flex-direction: column;
+  position: absolute;
+  top: 80px;
+  left: 50%;
   justify-content: space-between;
+  flex-direction: column;
   gap: 30px;
 }
 .input-messages {
-  position: absolute;
-  bottom: 6%;
-  right: 10%;
 }
